requirements.txt
dj-static==0.0.6
Django==1.10
django-filter==0.13.0
django-shortuuidfield==0.1.3
djangorestframework==3.4.3
djangorestframework-jwt==1.8.0
httpie==0.9.4
Markdown==2.6.6
psycopg2==2.6.2
Pygments==2.1.3
PyJWT==1.4.2
requests==2.11.0
shortuuid==0.4.3
six==1.10.0
static3==0.7.0
